Update decibel conversion functions

+++ src/core/tools/DecibelLinearizer.h
@@ -40,7 +40,7 @@ namespace SVS {
             return decibels > minusInfinityDb ? std::pow (10.0, decibels * .05) : .0;
         }
 
-        inline static double gainToDecibels(float gain, double minusInfinityDb = -96) {
+        inline static double gainToDecibels(double gain, double minusInfinityDb = -96) {
             return gain > .0f ? qMax (minusInfinityDb, std::log10 (gain) * 20.0) : minusInfinityDb;
         }
     };
